feat : 공모전 엔티티 빌더 추가
