--- src/main/java/CoCoNut_was/domains/project/entitiy/Category.java
@@ -1,4 +1,0 @@
-package CoCoNut_was.domains.project.entitiy;
-
-public enum Category {
-}
