+++ src/main/java/CoCoNut_was/domains/project/entity/Category.java
@@ -1,0 +1,4 @@
+package CoCoNut_was.domains.project.entity;
+
+public enum Category {
+}
