+++ src/main/java/CoCoNut_was/domains/project/entity/Project.java
@@ -1,6 +1,7 @@
-package CoCoNut_was.domains.project.entitiy;
+package CoCoNut_was.domains.project.entity;
 
 import jakarta.persistence.*;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -47,4 +48,20 @@ public class Project {
     public void onCreate() { // 생성일자 자동 생성
         this.createdAt = LocalDateTime.now();
     }
+
+    @Builder
+    public Project(String title, String merchantName, Category category, String description,
+                   String period, int rewardAmount, String summary, LocalDateTime createdAt,
+                   Status status) {
+        this.title = title;
+        this.merchantName = merchantName;
+        this.category = category;
+        this.description = description;
+        this.period = period;
+        this.rewardAmount = rewardAmount;
+        this.summary = summary;
+        this.createdAt = createdAt;
+        this.status = status;
+    }
+
 }
