+++ src/main/java/CoCoNut_was/domains/project/entity/Status.java
@@ -1,4 +1,4 @@
-package CoCoNut_was.domains.project.entitiy;
+package CoCoNut_was.domains.project.entity;
 
 public enum Status { // 진행중, 투표중, 종료
     IN_PROGRESS, VOTING, CLOSED
